util.h from Dave's DX9 port
DarkStarSword's merge notes:
This file was evidentially based on a different version of 3DMigoto to
the rest of the DX9 port. It does not appear to be based on a single
version as it includes changes from a mismatch of some commits, but
lacking others that precede them - but this seems to be the closest
singular point I can find in our history, so I started here and fixed up
as much as I could. This will need verification since it involved
significant alterations from Dave's work.

- Reverted irrelevant whitespace, comment and dead code changes that
  make the code review needlessly harder. While these are not bad, they
  should not be snuck in during a port, but rather should be submitted
  in a dedicated patch that makes no other functional changes.

- Reverted the disabling of all audible warnings. If the DX9 code is
  triggering these and for one reason or another cannot resolve them it
  should find another way to suppress them, such as #defining them to do
  nothing.

- Merged the two nearly identical BinaryToAsmText functions with an
  preprocessor directive around just the differing disassembler call.

- Reinstated d3d9.h and DirectXMath.h includes. Reverted struct float4
  back to use XMFLOAT4 from DirectXMath.h. Will need verification that
  this works on DX9 build as it may depend on which DirectX or Windows
  SDK is in use, and that was altered in the DX9 port.

- Added DX version checks around the save & restore output merger state
  helpers.

This still needs the hardcoded DX version define replaced with a project
setting before it can be merged into master, and the version checks
should not be using double negatives.