DX9: Standardise include paths across all configurations
We don't want the project configuration to be any more complex than it
needs to be and we want to keep as many settings the same across all
configurations as possible. IncludePaths is now the same as the DX11
IncludePaths, and AdditionalIncludePaths is as close as possible, plus
the DX SDK and removed the D3D_Shaders (which probably needs a global
fixup to rename its header file, separate to this work).