DX9/BinaryDecompiler: Fix compilation + reduce diff to upstream
The DX9 port of HLSLCrossCompiler to C++ did not allow for multiple
instances of a shader phase, which is understandable given that they are
irrelevant to DX9, but are required for some DX11 shaders, and as such
the DX11 backport used a slightly different data structure to facilitate
these. This commit adjusts the DX9 backport to use the same data
structure.

It also reduces the diff to MrShoor's fork of the upstream
HLSLCrossCompiler, replicating whitespace errors. Comments are added
where the code differed, either indicating that the difference is
3DMigoto specific, or querying it if the reason for the difference is
unknown.