DX9: Backport decompiler + assembler interface changes + drop zero marking mode
A while back we moved a bunch of parameters passed to the decompiler
into their own structure to simplify the code and minimise memory copies
when decompiling a shader. Backport this to DX9 so that it can compile.

The disassembler wrapper function has a few more arguments since the DX9
code forked off. Set these to false.

The zero marking mode was dropped due to being broken. Drop this from
DX9 as well.