DX9: Fix remaining build failures on 32bit
This fixes up a couple of remaining issues in util.h and the decompiler
to allow the 32bit build of the DX9 port to compile, and should
hopefully be roughly equivalent to the original code dump, but forward
ported onto a newer 3DMigoto code base. Note that other projects
(including DX11) remain broken by this port, and the 64bit DX9 build
throws numerous warnings and errors.

Compile tested only.