Disable building assembler separately
On the 64bit DX9 port the assembler fails to build as a standalone
project. We actually build it's source files in the DX9 and DX11
projects, so we don't need to build it standalone at all, so disable it
building in configuration manager.

With this change the DX9 port builds, albiet with many compiler
warnings.