Remove MIT license / Relicense 3DMigoto under the GPLv3
The DX9 port has included some code from GeDoSaTo, which is licensed
under the GPLv3. This is incompatible with the MIT license, and
therefore with the release of this port we can no longer distribute
3DMigoto under the MIT license.

Note that this situation differs from the inclusion of Nektra's
In-Process hooking library which also forces 3DMigoto to be licensed
under GPLv3 in some circumstances - in Nektra's case we keep our source
code separate allowing us to continue shipping our code under the MIT
license - it is only the 3DMigoto DLL that has been linked against the
Nektra library that becomes encumbered with the GPLv3 license (if the
individual building it lacked a commercial license from Nektra). In this
case however code has been copy + pasted from GeDoSaTo verbatim which
means that the GPLv3 now applies to our source code as well.

We could potentially keep the DX11 code under the MIT license and only
encumber the DX9 code, or we could strip the GeDoSaTo inclusions from
the DX9 code to allow it to be licensed under MIT, but I think that
would lead to further confusion and we have no specific reason to do so,
other than Chiri's original wish that the 3DMigoto code be available for
anyone to use with no restrictions.

Also add Dave to the AUTHORS.txt file.