DX9: Clean up QueryInterface_DXGI_Callback
- Fixes compile warning about using %x on a pointer
- Replaces verbose IID checks with IsEqualIID

This function could still use more of a cleanup. It appears to be
originally intended for the DXGI wrapper to call into, however a) we
don't use the DXGI wrapper, b) DXGI doesn't apply to DX9, c) We could
add a new exported function to facilitate communication between modules
that doesn't require the expensive GUID matches. It also appears to be
used internally to handle various QueryInterface calls, so we may not be
able to simply drop it. For now, just clean it up and fix the 64bit bug.