DX9: Fix crash when stereo is disabled
When stereo is disabled the mStereoHandle is not initialised leading to
a crash in nvapi. This doesn't happen in DX11 because the stereo texture
is not created if stereo is not initialised and that in turn leads us to
skip updating the stereo texture, but the DX9 port creates the stereo
texture unconditionally so it always attempts to update the stereo
texture.

This adds a check on the stereo handle to determine whether the stereo
texture should be updated and brings it in line with DX11, however it is
debatable whether we actually want this behaviour since the stereo
texture also contains some non-stereo things (resolution, tuning
variables) that probably shouldn't be sensitive to stereo... However it
is equally debatable whether those things actually belong in a *stereo*
texture at all since they are not stereo, and are largely left in for
backwards compatibility, and I'm not much inclined to fix a long
standing bug in long time deprecated functionality.