Fix crash in overlay if stereo enabled, but failed to create stereo handle
I've hit an edge case in DX9 when connected over Chrome Remote Desktop.
Stereo is enabled in the control panel and NvAPI_Stereo_IsEnabled
returns true, but NvAPI_Stereo_CreateHandleFromIUnknown fails with
NVAPI_STEREO_NOT_ENABLED. This results in a crash in the overlay as it
passes a NULL stereo handle to nvapi.

This adds an additional test in the overlay to make sure the stereo
handle is valid and treating it as a stereo disabled case otherwise.