DX9: Change ini filename to 3dmigoto9.ini
We don't want to use the same ini file as DX11 since there are some
small but significant differences between them that could cause problems
if both versions of 3DMigoto are installed to the same game folder for
one reason or another.

d3dx.ini was never a good name anyway since it doesn't indicate what
tool it belongs to, but we're stuck with that in DX11 for historical
reasons. Let's not make the same mistake a second time and call it
3dmigoto9.ini