DX9/Overlay: Remove weird direct destructor calls
Destructors should never be called like this. These font objects are
managed by std::unique_ptr, which will destruct them automatically when
they go out of scope, which will happen a moment after this as the
overlay's destructor returns. Delete the explicit destructor calls and
leave it to the smart pointer to manage them.

There are some additional direct destructor calls in the command list
that need to be looked at separately, but I will leave those for a
separate commit as they appear to have been explicitly changed from the
COM Release() calls in DX11 and I want to double check if there might be
a reason for the change first.