Add MIGOTO_DX=version defines
These are intended to be used in common code that is included in
multiple DirectX projects. This is not the method that is currently
being used by the DX9 port, but is intended to replace that. This:

- Does not require double negative tests for the common case of enabling
  code that only applies to a specific DX version
- Allows for minimum/maximum version checks for code that applies to
  e.g. DX10 and above

This adds the defines to the DX11 and DX10 projects - they will be added
to the DX9 project in a separate commit. The DX10 project is not
currently maintained, but these are added for future work to clearly
show our intent for how this should work.