Restore deletions from nvstereo.h
This restores a lot of the deleted code from nvstereo.h to support DX 9
and 10. It adds tags clearly showing which parts of the code have been
added/deleted for 3DMigoto, and overall (barring whitespace fixes) the
diff against the upstream nvstereo.h now look a little clearer. The
motivation behind this is to restore the DX9 parts of this file prior to
merging the DX9 port of 3DMigoto to more clearly see what that port has
changed.

It does make one significant alteration to the DX9 and DX10 code paths
vs upstream, which is to add a typedef for the context as the same type
as the device. This allows a single consistent UpdateStereoTexture API
to be used across all versions and won't break if someone should include
the header for multiple versions of DirectX. This will require some
alterations to the DX9 and DX10 ports, but is definitely worthwhile.

It defines the appropriate NO_STEREO_D3D macros based on MIGOTO_DX in
the project settings, if that itself is defined. It currently still
includes the d3d11_1.h file since as of this commit we only need to
worry about DX11 compatibility, and this may be adjusted when merging
the DX9 port, though that should not actually hurt to be included in any
project since it is not mutually exclusive with the other versions of
DirectX.

Tuning and resolution values are disabled if DX11 is disabled. These are
both broken in some way, inappropriate to place in a stereo texture and
deprecated and should not be used. They are left in for backwards
compatibility only, and we have no need to bring their brokenness into
DX9/10 so exclude them.