DX9: Drop D3D9Base namespace
This namespace around d3d9.h causes issues as most code outside the
DirectX9 3DMigoto project, including other DirectX headers and other
third party code does not expect it and will fail randomly. It prevents
us from being able to include both DX9 and DX11 headers together, for
example.

This commit is the result of the following commands to remove the
namespace, as well as manual removals of a few remaining instances and a
complete reversion of nvapi.h to the version from master:

    sed -i 's/D3D9Base::/::/g' DirectX9/*.{cpp,h}
    sed -i 's/D3D9Base:://g' *.{cpp,h} NVAPI/DllMain.cpp
    sed -i 's/^using namespace D3D9Base;$//g' *.{cpp,h} DirectX9/*.{cpp,h}

Within the DirectX 9 project the namespace is replaced with :: to
disambiguate it from any instances of the same names being used in the
wrapper namespace. Outside of the DX9 project no namespace is specified
at all since code that includes those files should not be aware of the
wrapper namespace. In the future the wrapper versions should be renamed
with a Hacker prefix to follow the same pattern as the DX11 project.