Assembler: Restrict new DX9 functions to DX9 project
This will only build the DX9 dis/assembler support when building the DX9
project, and will exclude it when building DX10+ since the June 2010
DirectX SDK that provides D3DX is not listed as a dependency of those
projects, and we do not wish to add it as one. We could potentially drop
D3DX entirely, since the dis/assemble functions it provides are also
available elsewhere, but IIRC there were some caveats with that.

This reverts the D3D_Shaders project file back to the version in master
- since the assembler source code is directly included in the target
projects (as opposed to statically linking it in) there is no need to
build this project separately and it does not need the DX9 dependencies
added. Arguably the DX9 functions should be moved out of the D3D_Shaders
project entirely since it does not use Flugan's assembler.

This commit drops the signature parser wrappers that are not applicable
to DX9, and throws an exception if the assembler fails to be caught by
the calling code, adding a try/catch block to one call site that was
missing it. The assembler currently does not pass warnings & errors back
to the caller, which needs to be added before release so we don't have a
repeat of Helix Mod.

With this change the DX11 project successfully compiles again.