TestShaders: Hook up HLSLCrossCompiler test cases to decompiler test suite
This adds all the HLSLCrossCompiler test cases to our HLSL decompiler
test suite. Note that a PASS here just means the shader was successfully
decompiled and recompiled and matched the current .chk files - it does
not mean that the decompiler output is necessarily correct to create a
shader with the same behaviour as the original.

The .chk files are just generated with the decompiler's current output
via --update-chk, and are used to identify any changes in output so that
we can validate by hand whether the change is intended or not.