TestShaders/BinaryDecompiler: Add HLSL source from HLSLCrossCompiler
This adds the source code for the various test cases from
HLSLCrossCompiler to our tree. These are just provided for reference to
see what the shader was originally, and are not (and cannot be) used by
the test suite.

Also move some test cases out of apps/shaders/shaders/ to just apps/shaders/