TestShaders/BinaryDecompiler: Add DX9 test cases from HLSLCrossCompiler
This adds test cases for ps_2_0, ps_3_0 and vs_2_0 from the
HLSLCrossCompiler project to give us a starting point to test the DX9
decompiler.

Adjusts the test suite to successfully locate the shader model from DX9
shader assembly output.

.chk files are not added as none of these decompiled successfully,
either producing a shader with an empty main function, or in the case of
BinaryDecompiler/ps3/constTexCoord.o outright crashing the decompiler.