DX9/Decompiler: Fix wrong buffer size passed to strcat_s
The decompiler is failing all DX9 test cases, due to the new strcat_s
added in applySwizzle not having enough buffer space to add the default
.xyzw swizzle. The issue is that the existing string length calculated
via strlen() is passed as the buffer size, which of course will never
have room to extend the string.

This switches the buffer size to use opcodeSize, as is already used
elsewhere in this function. This addition obsoletes the previous error
path in this function, though for safety that is left in since it
prevents the potential for random memory corruption, which would not be
fun.

.chk files are added (***dx9_wrong_history branch: only those relevant
to DX9***) since some of these now have some output in their bodies,
though other still lack any output and all have issues with
declarations.