DX9/Decompiler: Don't exclude shader models 1 or 2
The decompiler skips shader models 1 and 2 because sometimes these are
included as fallback shaders in DX11 for compatibility with older
hardware. These have been seen in some Unity games such as World of
Diving for example.

However, in DX9 we obviously don't want to skip these shaders, so make
the check dependent on not being a DX9 shader.

Updates vs2 .chk files. Other shader model 1/2 test cases are failing
for other reasons.