DX9/Decompiler: Rename 'inst' used for lookahead to 'instructions'
'inst' is not a good name since it is not immediately clear how it
differs from 'instr'. Rename it to 'instructions' to show that it is all
instructions and can be used to look ahead to upcoming instructions.

Refactor the instruction while loop to use this since since the backport
of BinaryDecompiler in master means both DX9 and DX11 should be using
the same data structures now.