DX9/Decompiler: Fix regression in dp3 instruction
This regression was detected by backporting the bug fixes necessary to
make the decompiler functional enough to generate regression test .chk
files to the dx9_wrong_history branch, then comparing the .chk files in
both branches. The DX11 code path for the DP3 instruction was not meant
to be in an else block, as it is used as a fallback for DX9 as well when
the lookahead code does not match instructions corresponding to the HLSL
reflect() function.