DX9/Decompiler: Add range checks to all lookahead paths
The DX9 port includes lookahead code in the decompiler that aims replace
sequences of multiple instructions with a single HLSL intrinsic function
call. If the first instruction in such a sequence appears near the end
of the shader the lookahead code will attempt to look beyond the end of
the shader's instruction stream, triggering a debug assertion in debug
builds and potentially crashing in release builds.

This patch adds appropriate range checking to all these paths to ensure
they won't lookahead when there aren't enough instructions remaining to
match the pattern they are looking for.

No changes to any regression test output, however the regression tests
no longer trigger debug assertions when using a debug build of
cmd_Decompiler.