CopyToGames: Use CopyToGamesLocal.bat if it exists
This allows other developers to make use the CopyToGames mechanism to
speed up the testing process. If a file named CopyToGamesLocal.bat
exists it will be called from CopyToGames.bat.

Bo3b's original CopyToGames.bat is left in place and can serve as a
template for creating a CopyToGamesLocal.bat.