Solution File: Re-enable DX11 and DXTK project builds
Both DX9 and DX11 projects can now be built from the same tree, so
enable both in the solution configuration, as well as the DirectXTK
project that had been disabled in some configurations for an unspecified
reason.

The solution file diff against master now boils down to enabling the DX9
project and disabling Flugan's assembler built as a standalone project.