Refactor software cursor into separate file
This code does not belong in IniHandler.cpp as it is neither anything to
do with ini handling, and is substantial enough to warrant being placed
in its own file.

Renamed the two similar yet subtly different InstallHook functions to
signify the difference between them, and removed the useless
LogInfoIsSafe boolean that was supposed to differentiate them back when
they shared a common code path.