DX9: Disable LoadLibrary redirect if 3DMigoto's d3d11.dll is loaded
If both 3DMigoto's d3d9.dll and d3d11.dll are loaded they will both
install a LoadLibrary hook, and both attempt to redirect nvapi to the
game directory. The result of this is that when the nvapi wrapper
attempts to load the system nvapi, the first 3DMigoto wrapper will
translate the special "original_nvapi" string back to the system path,
and the second wrapper will redirect that back to the game directory.
The nvapi wrapper will be given a reference to itself, and will end up
recursively calling itself until the stack runs out and crashes.

Since we can't be sure that end users will remove/update the d3d11.dll
we have to assume that an old version of it might be present, and must
cope with this in the d3d9.dll regardless of which is loaded first.

This commit adds a check in the d3d9.dll for a 3DMigoto d3d11.dll. The
check is done by the version tag embedded in the DLL, which has been
present ever since 3DMigoto v1.0. The code is based on the version check
in the 3DMigoto loader executable.