DX9: Fix crash due to texture bound to multiple slots being released prematurely
In DirectX9 (and 10) an object's refcount can drop to zero while it is
still alive by merit of being bound to the pipeline. The DX9 port
includes code to handle this, by keeping track of which objects are
bound separate from their refcount, however it failed to allow for
objects to be bound to multiple slots on the pipeline simultaneously,
allowing for situations where an object would be deleted prematurely
after it had been unbound from one slot in the pipeline, while still
bound to another.

This pattern was seen in the main menu of Miasmata leading to a crash:

    IDirect3DSurface9::Bound stage=3 handle=11BBB660 counter=2 this=10EEB228
    IDirect3DSurface9::Bound stage=2 handle=11BBB660 counter=2 this=10EEB228
    IDirect3DSurface9::Unbound stage=3 handle=11BBB660 counter=2 this=10EEB228
    IDirect3DSurface9::Bound stage=1 handle=11BBB660 counter=2 this=10EEB228
    IDirect3DSurface9::Unbound stage=2 handle=11BBB660 counter=2 this=10EEB228
    IDirect3DTexture9::Release handle=11BBB660, counter=0, internal counter = 0
    IDirect3DTexture9::Delete handle=11BBB660 counter=0 this=10EEB228
      deleting self
    IDirect3DSurface9::Unbound stage=1 handle=00000000 counter=0 this=10EEB228
    IDirect3DTexture9::Delete handle=00000000 counter=0 this=10EEB228
      deleting self

This commit changes the bound flag on texture objects from a boolean to
a set, allowing it to keep track of which slots it is bound to, and only
delete itself once it has been fully released and unbound from all slots.

For now, this is the only object with this change - it is likely that
other objects may need the same treatment.