Experimental Mutex around each Context->Begin/End.
Does not work to solve the crash in The Outer Worlds while running under Win10 store with XBox Game Pass for PC.

Mutex itself seems to be working correctly, does not hang or show other deadlock issues.

This commit is part of a new exp_ContextMutex branch to capture this work in the unlikely case it has some value down the road.