Disable building DX9 as part of solution
The DX9 port brings in an extra dependency (The old June 2010 DirectX
SDK, specifically for D3DX, which was deprecated and dropped when the DX
SDK was merged into the Windows SDK), which means it will fail to build
for existing developers.

Given that the DX9 port is not ready, there's not much point hassling
everyone to install it just yet, so disable it in the solution
configuration - the DX9 project can always be built directly as needed.