DX9: Fix compile failure with VS2017
The DX9 port has templatized the expression parser, and the VS2017
toolchain seems to be unable to resolve one of the templated classes to
find a typedef, possibly due to a compiler bug. This typedef does not
actually depend on the template argument, so replace it with the full
canonical non-typedefed type.

The VS2017 toolchain produces a myriad of compiler warnings in the DX9
project that VS2012 did not. At a cursory glance many of these do look
like legitimate bugs that will need to be ironed out, however the
project does build and successfully loads into Miasmata.