Move DirectX11 project to top of sln file
When someone opens the solution up for the first time in Visual Studio,
the first project in the sln file will be set as default, which is
currently the DirectX9 project. To help assist new developers in
building the correct target, move the DirectX11 project to the top of
the list so that it gets set as default.

Unfortunately the rest of the build configuration defaults seem to be
set alphabetically, so Debug/Win32 will be set instead of the preferable
Zip Release/x64.