Switch all tabs to 4 spaces.
This will be a hugely noisy change, some 68,000 items changed.  Any comparisons to previous code will be hard to see after this, which is why we are doing this as early as possible.

Done using global find/replace via \t regex. Not using code reformatters, so we can be sure that this commit is solely white space change tabs to 4 spaces.