Second stage of convert tab to 4 spaces.
Using offline notepad++ to convert all .cpp, .h, .hpp files in the directory.  This thus includes all DX9 and DX10 files.

Some files were reverted for DirectXTK, crc32, pcr2, as those are code we don't want to change.

We also did not change BinaryDecompiler as the JamesJones code, as we might want to import a new version later and not want whitespace noise.

Project still compiles and builds.