Add using namespace std to .cpp files.
Without an accidentally included using namespace, we now need to explicitly specify we are using namespace std; in all the cpp files that using std functions.

It's OK to include in .cpp files because they are not included by other compile units.

It would be possible to specify just the component used, like using vector, or using wstring, but just using the whole namespace here seems more consistent and simpler.