Change all log calls to LOG_INFO style.
This changes all the logging macros in log.h to use LOG_INFO and LOG_DEBUG macro style, from the previous LogInfo, LogDebug.

This was done using text substitution with match-word, so that we can hit everything.  The rename style refactor fails on macros for whatever reason.

Compiles without errors on DX11.

Other targets like DX10 and DX9 are also being changed, but we are not trying to keep those running.