Change touch_dir macro to function.
Making it match the touch_file semantic, and removing a macro is always good.

This commit changes the last of macros that we wanted to make consistent with naming guideline.

Searched for "#define [a-z]" with match case enabled, found remaining macros with lower case.

Not fixing macros manually in DX9 or DX10 project, but any matches were automatically changed.