LockResourceCreationMode -> LOCK_RESOURCE_CREATION_MODE.
As a macro, this needs to be ALL_CAPS, so changing the defintion and all uses for it.

For this LockResource mechanism, it seems like it would be better to always be the Function, and Line macro pretty version instead of making a distinction, but I won't change this now.