EnterCriticalSectionPretty -> ENTER_CRITICAL_SECTION
And also add a LEAVE_CRITICAL_SECTION to match any use of this macro.  This is the pretty version with Function and Line number, so any use of these macros is using the pretty variant.

And will be visually consistent with the ENTER and LEAVE, which will more clearly define critical sections.

DX9 and DX10 projects were skipped for these.

Full batch rebuild of all 6 targets was successful.  Should be no change in runtime behavior.