Rename function parameters from mOrig naming.
These were all specified using the mHackerDevice naming, which was wrong to begin with because these are function parameters not member variables.

There was also a mix of *1 types, and the defintions did not always match the implementations.

This commit straightens all that out to make it consistent and match our style guidelines, with the input parameters in snake_case.