These are changes to DX9 subproject for renaming.
This is an extensive set of automatic changes to the DX9 subproject, as part of Resharper renamings.  This were also done automatically, but not checked.

DX9 will not compile at present, but these changes might be useful later on.

Doing both the DX9 and DX11 at the same time is too much work, so DX9 will need to revisited.