Update D3D11Wrapper.cpp for style.
This fixes all the obvious style inconsistencies in the file and the header, to match our new style guidelines of lower_snake_case for global functions.

There are a lot of files that are touched by changing routines here, like NvAPIOverride->nvapi_override.

Some changes are recommended by Resharper, but don't really make sense because they are Microsoft function names that should be kept consistent with Microsoft naming.