Fix clang compilation errors.
For using clang-tidy, we need to be able to compile using the clang compiler as the front end.  This fixes the compile errors that clang sees that msvc does not.

Fixes some include path errors, including <> that should be "", and some missing includes that somehow work in msvc.  Adds specific std library includes like <vcruntime_typeinfo.h> for rtti info.

Removes use of unsupported 'default' in template.

Removes use of unsupported 'for each' in Overlay.

This now compiles with clang, and clang-tidy can find and fix errors.  Still also compiles under normal msbuild.