Full reformat of Util.h and Util.cpp.
This uses the clang-format to reformat both these files to match our expected style.  It's a lot of changes.

All pointer references and definitions are now left justified. Since we have not been fully consistent, some moved, some did not.

Indenting and bracket use is not fully consistent with an emphasis on white space to show the surrounding block. And all lines are distinct without doubling up lines of code.  All for/while/if blocks are now Allman style with containing brackets.

switch statements are now more indented to more clearly show the parts and the actual cases.

Lines are no longer wrapped for defintions or usage, as we set it to no wrapping.

Only white space changes, so compiles in both.