Move all routines out of util.h
This is bad form to have so much actual code in the .h file, so let's move it out into the .cpp file where it belongs, and leave the type defintions there.

This is fairly involved with some risk.  Lots of bad #ifdef traps and being declared static.  Removed all the use of static.

Templates have to stay because of how primitive they are.

All 6 build targets still compile.