Update annotations on HackerContext.cpp
These /* annotation */ are useless and stupid, because it's already obvious.  They also break the clang-format, so we are removing them all from the cpp file.  We'll leave them in the .hpp file so that the signature can look the same as the d3d11.h file.

Also remove all the THIS_ macros, which don't do anything, but also break the automatic formatting.  Not necessary as a reminder.

Fixes all the _In_ type annotations to match the .hpp file which has an updated variant of the SAL 2.0.  These being mismatched cause clang-tidy to get lost and generate too many errors.  Replaced via text search/replace.