Rename all parameters of Hunting using clang-tidy.
Bit of an experiment here, using clang-tidy with "readability-identifier-naming.ParameterCase, value: lower_case"

That finds all params with alternate style, and tags them with warnings.  Adding the --fix to the command line fixes all of them automatically.

That left behind some pData, and ppShader type vars, renamed as p_data, and pp_shader, so we also did a regex search to find those and remove the leading pp_ and p_.