Fix all parameters for style in HackerContext.
This updates the parameters in object methods to match guidelines.  It does not change the ones defined by the DX11 wrapper functions, only the standalone functions.

There are some routines that take parameters like StartSlot, NumViews, pShader and so on, that are not changed, as these are direct copies of the DX11 formed parameters.

Still compiles.