Do full clang-format on HackerContext.
This will change nearly all whitespace in the HackerContext.cpp and HackerContext.hpp files.

The reformat follows the rule specified in the included .clang-format as of this commit.  Since this is automatically applied by VS upon every paste, closing bracket or save, we should be able to keep this formatting indefinitely.

Primary changes:
Pointer specification is left justified
All assignments are aligned for visual simplicity.
All declarations are aligned for visual simplicty.
All input parameters are on separate lines in cpp file, on a single line in hpp file.
All if/then/while/for are on separate lines.
Brace style forced to Allman.
Indenting in blocks is enforced.
No line limit, so no wrapping.

This should be the exact format we expect to use going forward.

All files compile and build with all 6 targets.