Remove binary_to_asm_text from DisassembleFlugan.
This fixes a compile error in cmd_Decompiler, that was introduced when moving the Util code out of .h and into the .cpp.

Because of really complicated include paths, there is no good way to include the binary_to_asm_text routine, without having to include Overlay and HookedContext and HookedDevice, and DirectXTK and so on.

This just copies the binary_to_asm_text directly to cmd_Decompiler.cpp and skips all that.