Remove all SAL notations from HackerContext.cpp
We'll leave them in the .hpp file as reference and as the copy of the original DX11 calls, but are removing them all from the actual implementations in the .cpp file.

This makes for much more readable code, without all the macros, especially after they are aligned.

This also restores the original formatting to the .hpp file, so that the function definitions will have the exact orginal formatting as the DX11 functions. Their formatting is a bit more readable when the SAL macros remain.