Remove all SAL macros from HackerDevice.
This includes all the different variants in both our functions in .hpp, as well as all variants in the .cpp file.

We leave all the macros in the .hpp file for override versions that match Microsoft functions.  We also mark that section to not be clang-format with the // clang-format off comment.  And restore all the formatting so that it matches the original DX11 format.

Full reformat with ctrl-k, ctrl-d

Still compiles.