Rename all variables and functions in HackerDevice.
This uses the included .clang-tidy file to rename all the functions and variables in the file, to match the layout we decided for naming style.

This includes every facet of the language now, and using this .clang-tidy and ClangPowerTools, it can do an automated fix for the naming.

The one glitch is that ParameterCase is commented out, because otherwise it will also alter every DX11 function override, and we want to keep those as is.  So for our functions and methods, they were done manually using Resharper.

All 6 build targets compile and link.