Rename vars and functions in HookedDXGI.
This is the full rename of the variables, while attempting to avoid changing named vars that match DX11 functions hooked or wrapped.

The technique used here is with the .clang-tidy file, where we can specify a regex of names to avoid, so I've set it up with all the candidates here as:
value: "(This|pDevice|pDesc|pWindow|hWnd|pFullscreenDesc|pRestrictToOutput|ppSwapChain|pSwapChain.?|ppFactory.?|pFactory.?|Flags)"

So all those var names are ignored by clang-tidy, and thus not changed.  This leaves some vars unchanged that we actually want changed in normal functions, but those are fixed manually as a smaller set, and often have a conflicting pDesc type form anyway.

We can leave the .clang-tidy this way, and it will avoid marking any of these as misnamed.

All 6 build targets compile.