Replace every faulty {0} initializer, solution wide.
These initializers were wrong.  In C++ if you specify a parameter in the list, it becomes a specific initializer for a given set of parameters, not clearing the entire struct to zero as we want.

This commit replaces every {0} and { 0 } with {}, which will zero initialize the entire struct.

Batch build of all 6 targets is successful.