Run modernize-deprecated-headers against full solution.
This changes all C style includes like math.h or stdlib.h into the C++ format of <cmath> and <stdlib>.

Moved the _USE_MATH_DEFINES into the solution PreProcessor defs so that we can find M_PI.

We restore files in crc32 and BinaryDecompiler as those are included projects that we are not updating.

All 6 build targets successfully build.