Rename lower case files to upper case.
This is all inconsistent naming for the files, with some lowercase, and some UpperCase.  This commit unifies all the project files for DirectX11 to be UpperCase as more like Objects.

So cursor->Cursor, globals->Globals, lock->Lock.  The references in the #includes are also updated to reflect the naming, even as windows in not case sensitive, but to be consistent.

The upper level files are left lower case, like log.h, util.h, as those are used in other projects here, and it's a nice visual difference to show those.

As part of this cleanup, all backslash are converted to forward slash.

All 6 build targets remain successful.