Update and sort every include file.
This sorts every set of include files at the top of every file, and puts them into a consistent order of
1) reference.h for a .cpp if valid.
2) All local .h files from our project
3) All global includes like <system>

This now follows the approach of local to global for all include paths. This is the recommendation from StackOverflow and other sites, that this sort order will catch errors faster than any other approach.

There are a couple of strange glitches- the ResourceHash.cpp needs a special sequence because DirectXTK doesn't directly support our use of embedded .h files.

And the NVStereo project needs to have <d3d9.h> included first, so that it can define _D3D9_H_ that nvapi.h needs in order to define all functions we need to override.

For the NVStereo dll, this also removes the main.h file, and just puts those includes directly into Main.dll, as it's not used anywhere else.

All 6 build targets are successful.