Unify all include file naming to match originals.
Lots of naming errors that don't match the actual file names.  Used Resharper to fix all of these automatically so that the case matches.  Not super important as Windows is case-insensitive, but there was a lot of ALL caps and other strangeness.

Changed uses of <pcre2.h> to "pcre2.h" because these are local to our project, not system includes.

Removes all unused #include files as well.  Looks like a second pass was necessary to get them all.

All 6 build targets are sucessful.