Clean up the solution explorer.
Move files in the DirextXDX11 that are shared and lowercase into a Shared filter folder, as these are shared with other projects in solution.

We have a lot of files, and this shortens the list a bit.

Rename some files in the explorer so that they have the same case as the files on disk.