Rename all variables in Assembler project.
The variables are all renamed as the clang-tidy files specifies, with a couple of added skipped vars that seemed better to leave all caps inline.

The function names are skipped here, because of lots of obscure acronyms in their names that otherwise get flattened out and less clear.

This project does not change as much, so I'm using a lighter hand on this code, only trying to avoid the previous mismatch here with half vars being lower_case, and half being camelCase.

Full project still compiles and builds all 6 targets.